HW1/requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.3.0
Pillow>=8.0.0
scikit-learn>=0.24.0
pandas>=1.3.0
tqdm>=4.60.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.10.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.800

# Optional Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.0.0